+++ src/strenc/b64.cpp
@@ -22,7 +22,7 @@ std::string b64_encode(std::string in_string, int padded_len) {
     );
     // return output string
     std::string enc_string = encoded.substr(0, encoded.length() - 1); //a null character we don't want gets included
-    if (padded_len != -1 && padded_len > enc_string.length()) {
+    if (padded_len != -1 && (size_t) padded_len > enc_string.length()) {
         enc_string += std::string(padded_len - enc_string.length(), '=');
     }
     return enc_string;
